requirements.txt
click==8.1.7
Flask==3.0.0
Flask-Cors==3.0.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
six==1.16.0
werkzeug>=3.0.0
zipp==3.17.0
tensorflow>=2.13.1
flask_sqlalchemy>=3.1.1
google-cloud-texttospeech
google-cloud-vision